+++ client/utils/redux-form/normalizers/date/index.js
@@ -1,0 +1,2 @@
+export { default as mmyy } from './mmyy'
+export { default as ddmmyyyy } from './ddmmyyyy'
