+++ client/utils/redux-form/normalizers/index.js
@@ -1,4 +1,6 @@
+import * as dateNormalizers from './date'
+export const date = dateNormalizers
+
 export { default as creditcard } from './creditcard'
-export { default as date } from './date'
 export { default as number } from './number'
 export { default as string } from './string'
